fix: renamed otp & sms files
