+++ app/src/application/otp/otpMessageBuilder.js
@@ -1,0 +1,3 @@
+module.exports.otpMsgITA = 'è il tuo codice verifica del chatbot richiesta informazioni AcegasApsAmga.'
+
+module.exports.buildOtpMessage = (code) => `${code} ${this.otpMsgITA}`
