--- app/src/application/smsMessageBuilder.js
@@ -1,3 +1,0 @@
-module.exports.smsMsgITA = 'è il tuo codice verifica del chatbot richiesta informazioni AcegasApsAmga.'
-
-module.exports.buildSmsMessage = (code) => `${code} ${this.smsMsgITA}`
